+++ src/components/schedule.rs
@@ -1,18 +1,24 @@
 use std::cmp::Ordering;
 
-use crate::app::HomeOrAway;
-use crate::constants::TEAM_NAMES;
+use crate::app::{AppSettings, HomeOrAway};
+use crate::components::constants::TEAM_NAMES;
 use chrono::{DateTime, NaiveDate, ParseError, Utc};
 use chrono_tz::America::Los_Angeles;
+use chrono_tz::Tz;
 use core::option::Option::{None, Some};
 use mlb_api::schedule::{Game, ScheduleResponse};
 use tui::widgets::TableState;
+
+// TODO configurable timezone
+const TIMEZONE: Tz = Los_Angeles;
 
 /// ScheduleState is used to render the schedule as a `tui-rs` table.
 pub struct ScheduleState {
     pub state: TableState,
     pub schedule: Vec<ScheduleRow>,
     pub date: NaiveDate,
+    /// Used for selecting the date with arrow keys.
+    pub selection_offset: i64,
 }
 
 /// The information needed to create a single row in a table.
@@ -29,21 +35,23 @@ pub struct ScheduleRow {
 
 impl Default for ScheduleState {
     fn default() -> Self {
-        let date = Utc::now().with_timezone(&Los_Angeles).date_naive();
+        let date = Utc::now().with_timezone(&TIMEZONE).date_naive();
         ScheduleState {
             state: TableState::default(),
             schedule: vec![],
             date,
+            selection_offset: 0,
         }
     }
 }
 
 impl ScheduleState {
-    pub fn from_schedule(schedule: &ScheduleResponse) -> Self {
+    pub fn from_schedule(settings: &AppSettings, schedule: &ScheduleResponse) -> Self {
         let mut ss = ScheduleState {
             state: TableState::default(),
-            schedule: ScheduleRow::create_table(schedule),
+            schedule: ScheduleRow::create_table(settings, schedule),
             date: ScheduleRow::get_date_from_schedule(schedule),
+            selection_offset: 0,
         };
         ss.state.select(Some(0));
         ss
@@ -51,30 +59,37 @@ impl ScheduleState {
 
     /// Update the data from the API. It is assumed that the date is already updated, aka don't use
     /// a random date without first setting the `date` field. Use `set_date_from_input` for this.
-    pub fn update(&mut self, schedule: &ScheduleResponse) {
-        self.schedule = ScheduleRow::create_table(schedule);
+    pub fn update(&mut self, settings: &AppSettings, schedule: &ScheduleResponse) {
+        self.schedule = ScheduleRow::create_table(settings, schedule);
     }
 
     /// Set the date from the input string from the date picker.
     pub fn set_date_from_input(&mut self, date: String) -> Result<(), ParseError> {
         self.date = match date.as_str() {
-            "today" => {
-                // TODO configurable timezone
-                Utc::now().with_timezone(&Los_Angeles).date_naive()
-            }
+            "today" => Utc::now().with_timezone(&TIMEZONE).date_naive(),
             _ => NaiveDate::parse_from_str(date.as_str(), "%Y-%m-%d")?,
         };
         self.state.select(Some(0));
         Ok(())
     }
 
+    /// Set the date using Left/Right arrow keys to move a single day at a time.
+    pub fn set_date_with_arrows(&mut self, forward: bool) -> NaiveDate {
+        match forward {
+            true => self.selection_offset += 1,
+            false => self.selection_offset -= 1,
+        }
+        Utc::now().with_timezone(&TIMEZONE).date_naive()
+            + chrono::Duration::days(self.selection_offset)
+    }
+
     /// Return the `game_id` of the row that is selected.
     pub fn get_selected_game(&self) -> u64 {
-        match self.schedule.get(
-            self.state
-                .selected()
-                .expect("there is always a selected game"),
-        ) {
+        let idx = match self.state.selected() {
+            Some(s) => s,
+            None => return 0,
+        };
+        match self.schedule.get(idx) {
             Some(s) => s.game_id,
             _ => 0,
         }
@@ -160,12 +175,20 @@ impl ScheduleRow {
     }
 
     /// Transform the data from the API into a vector of ScheduleRows.
-    fn create_table(schedule: &ScheduleResponse) -> Vec<Self> {
+    fn create_table(settings: &AppSettings, schedule: &ScheduleResponse) -> Vec<Self> {
         let mut todays_games: Vec<ScheduleRow> = Vec::with_capacity(schedule.dates.len());
         if let Some(games) = &schedule.dates.get(0) {
+            let favorite = settings
+                .favorite_team
+                .clone()
+                .unwrap_or_else(|| "na".to_string());
             for game in &games.games {
                 for g in game {
-                    todays_games.push(ScheduleRow::create_matchup(g));
+                    if g.teams.home.team.name == favorite || g.teams.away.team.name == favorite {
+                        todays_games.insert(0, ScheduleRow::create_matchup(g));
+                    } else {
+                        todays_games.push(ScheduleRow::create_matchup(g));
+                    }
                 }
             }
         }