+++ src/draw.rs
@@ -6,7 +6,7 @@ use tui::widgets::{Block, BorderType, Borders, Clear, Paragraph, Tabs};
 use tui::{Frame, Terminal};
 
 use crate::app::{App, DebugState, MenuItem};
-use crate::debug::DebugInfo;
+use crate::components::debug::DebugInfo;
 use crate::ui::at_bat::AtBatWidget;
 use crate::ui::boxscore::TeamBatterBoxscoreWidget;
 use crate::ui::help::{HelpWidget, DOCS_LEN};
@@ -33,9 +33,9 @@ where
 
     terminal
         .draw(|f| {
-            main_layout.update(f.size(), app.full_screen);
+            main_layout.update(f.size(), app.settings.full_screen);
 
-            if !app.full_screen {
+            if !app.settings.full_screen {
                 draw_tabs(f, &main_layout.top_bar, app);
             }
 
@@ -197,7 +197,7 @@ where
         false => Style::default().fg(Color::Red),
     };
     let block = Block::default()
-        .title("Enter a date in YYYY-MM-DD form")
+        .title("Enter a date (YYYY-MM-DD) or use arrow keys")
         .borders(Borders::ALL)
         .border_type(BorderType::Rounded)
         .border_style(border);