+++ src/event.rs
@@ -1,5 +1,5 @@
 use crate::app::{HomeOrAway, MenuItem};
-use crate::stats::TeamOrPlayer;
+use crate::components::stats::TeamOrPlayer;
 use crate::{app, cleanup_terminal};
 use crossbeam_channel::Sender;
 use crossterm::event::KeyCode::Char;
@@ -55,6 +55,16 @@ pub fn handle_key_bindings(
             } else {
                 app.state.date_input.is_valid = false;
             }
+        }
+        (MenuItem::DatePicker, KeyCode::Right) => {
+            let date = app.state.schedule.set_date_with_arrows(true);
+            app.state.date_input.text.clear();
+            app.state.date_input.text.push_str(&date.to_string());
+        }
+        (MenuItem::DatePicker, KeyCode::Left) => {
+            let date = app.state.schedule.set_date_with_arrows(false);
+            app.state.date_input.text.clear();
+            app.state.date_input.text.push_str(&date.to_string());
         }
         (MenuItem::DatePicker, KeyCode::Esc) => {
             app.state.date_input.text.clear();